+++ mysql-haskell-openssl/ChangeLog.md
@@ -1,5 +1,10 @@
 # Revision history for mysql-haskell-openssl
 
+## 0.8.0.0  -- 2016-11-09
+
+* Add `ciCharset` field to support `utf8mb4` charset.
+* Add `BitMap` field to `COM_STMT_EXECUTE`, and [#8](https://github.com/winterland1989/mysql-haskell/pull/8) by [alexbiehl](https://github.com/alexbiehl).
+
 ## 0.7.0.0  -- 2016-11-09
 
 * Split from mysql-haskell.